04_MCP/01_sqlite/requirements.txt
requests
langchain
langgraph
langchain-mcp-adapters
langchain-openai
fast-agent-mcp
fastmcp
questionary
